add test on async delete split client tests into two files
